+++ concent_api/core/tests/test_upload_finished_task.py
@@ -3,14 +3,15 @@ import mock
 from django.conf import settings
 from freezegun import freeze_time
 
+from common.constants import ConcentUseCase
+from common.helpers import get_current_utc_timestamp
+from common.helpers import parse_datetime_to_timestamp
+from common.helpers import parse_timestamp_to_utc_datetime
 from core.message_handlers import store_subtask
 from core.models import PendingResponse
 from core.models import Subtask
 from core.tasks import upload_finished
 from core.tests.utils import ConcentIntegrationTestCase
-from common.helpers import get_current_utc_timestamp
-from common.helpers import parse_datetime_to_timestamp
-from common.helpers import parse_timestamp_to_utc_datetime
 
 
 class UploadFinishedTaskTest(ConcentIntegrationTestCase):
@@ -83,7 +84,7 @@ class UploadFinishedTaskTest(ConcentIntegrationTestCase):
     def test_that_scheduling_task_for_subtask_after_deadline_should_process_timeout(self):
         datetime = parse_timestamp_to_utc_datetime(get_current_utc_timestamp() + settings.CONCENT_MESSAGING_TIME + 1)
         with freeze_time(datetime):
-            with mock.patch('core.tasks.payments_service.make_force_payment_to_provider', autospec=True) as payment_function_mock:
+            with mock.patch('core.payments.bankster.finalize_payment', autospec=True) as finalize_payment:
                 upload_finished(self.subtask.subtask_id)  # pylint: disable=no-value-for-parameter
 
         self.subtask.refresh_from_db()
@@ -92,9 +93,10 @@ class UploadFinishedTaskTest(ConcentIntegrationTestCase):
         self.assertEqual(PendingResponse.objects.count(), 2)
         self.assertTrue(PendingResponse.objects.filter(client=self.subtask.provider).exists())
         self.assertTrue(PendingResponse.objects.filter(client=self.subtask.requestor).exists())
-        payment_function_mock.assert_called_once_with(
-            requestor_eth_address=self.task_to_compute.requestor_ethereum_address,
-            provider_eth_address=self.task_to_compute.provider_ethereum_address,
-            value=self.task_to_compute.price,
-            payment_ts=parse_datetime_to_timestamp(datetime),
+        finalize_payment.assert_called_once_with(
+            subtask_id=self.subtask.subtask_id,
+            concent_use_case=ConcentUseCase.ADDITIONAL_VERIFICATION,
+            requestor_ethereum_address=self.task_to_compute.requestor_ethereum_address,
+            provider_ethereum_address=self.task_to_compute.provider_ethereum_address,
+            subtask_cost=self.task_to_compute.price,
         )